#1439: split pkg/dataplane/userspace/snapshot.go into 14 sibling files
Pure code motion within package userspace. Splits the 2,404-LOC
snapshot.go monolith (63 functions) into 14 aspect-named sibling
.go files, all in the same package userspace:

  builder.go     orchestrator + content hash + map pins  (4 fns)
  interfaces.go  interface + link + address helpers     (16 fns)
  fabric.go      fabric snapshots                        (2 fns)
  mirrors.go     port-mirroring snapshots                (2 fns)
  tunnels.go     tunnel endpoint snapshots               (1 fn)
  zones.go       zone snapshots + interface->zone map    (2 fns)
  routes.go      route snapshots + helpers               (4 fns)
  neighbors.go   neighbor snapshots + monitored ifs      (6 fns)
  nat.go         source/static/dest/NAT64/Nptv6 NAT      (8 fns)
  screens.go     screen profiles + SYN cookie keys       (5 fns)
  flow.go        flow + flow export snapshots            (2 fns)
  filters.go     firewall filters + policers             (4 fns)
  cos.go         class-of-service builder                (1 fn)
  policies.go    security policies + scheduler-state     (6 fns)

Total: 4+16+2+2+1+2+4+6+8+5+2+4+1+6 = 63 functions, matching the
grep oracle exactly. snapshot.go is deleted.

Bare aspect names per Wave 4 standing rule (no snapshot_ prefix).
Public API preserved verbatim: UserspaceBoundLinuxInterfaces moves
to interfaces.go; MonitoredInterfaceLinkIndexes moves to
neighbors.go. Daemon imports unchanged.

filterPublishableNeighbors stays in manager.go (cross-file,
same-package call from builder.go's snapshotContentHash).

Methodology: triple-review plan-review iterated 3 rounds with
Codex + Gemini + Antigravity to converge on the 14-file layout
(plan history preserved at docs/pr/1439-snapshot-builders/plan.md).
Code-review iterated 2 rounds with all four reviewers (Codex,
Gemini, Antigravity, Copilot) reaching MERGE-READY at HEAD.

This commit is a clean rebase carry-forward on top of fresh master
(b84f46280) after GitHub's merge planner declined the prior
multi-commit branch head. Function bodies verified byte-equivalent
to pre-split snapshot.go (/tmp/verify_bodies.py: 63/63 identical).

Gates passed:
- go build ./... clean (first try after goimports).
- go vet ./pkg/dataplane/userspace/... clean (only pre-existing
  warnings in cli.go and daemon_flow.go on master, unrelated).
- go test ./pkg/dataplane/userspace/... -count=1 PASS.
- Full go test ./... PASS.
- 5x flake check on TestFilterPublishable and
  TestUserspaceBoundLinuxInterfaces: 5/5 ok.
- byte-equivalence: all 63 functions identical to pre-split.

Closes #1439.
