+++ docs/pr/1439-snapshot-builders/reviewer-ids.md
@@ -1,0 +1,30 @@
+# #1439 reviewer task IDs
+
+Long-running agents lose Codex/Gemini session state; task IDs are
+preserved here so continuations can refetch by id.
+
+## Plan review round 1
+
+- Codex: task-mpn5xksb-unj08f
+- Gemini: task-mpn5y4dq-yydg6k
+- Antigravity: review-mpn5yls7-nbyi2u
+
+## Plan review round 2
+
+- Codex: task-mpn67k1p-5jeu4s
+- Gemini: task-mpn680k0-o6i3zb
+- Antigravity: review-mpn6800x-lpplct
+
+## Plan review round 3
+
+- Codex: task-mpn6fpvn-p002vc
+- Gemini: task-mpn6g24b-6fmq6w
+- Antigravity: review-mpn6g0sn-4983tc
+
+## Code review round 1
+
+- PR: #1592
+- Codex: task-mpn6obqh-9e2c7p
+- Gemini: task-mpn6oon6-wy6tv0
+- Antigravity: review-mpn6oo4d-hfxgxr
+- Copilot: requested via gh pr edit --add-reviewer Copilot
