+++ pkg/dataplane/userspace/interfaces.go
@@ -1,0 +1,465 @@
+package userspace
+
+import (
+	"fmt"
+	"net"
+	"os"
+	"path/filepath"
+	"sort"
+	"strings"
+
+	"github.com/psaab/xpf/pkg/config"
+	"github.com/vishvananda/netlink"
+)
+
+const (
+	// Keep logical-only synthetic ifindexes in a high private range so
+	// they do not collide with kernel-assigned ifindexes in practical
+	// deployments, while remaining positive int32 values for protocol
+	// compatibility.
+	syntheticInterfaceIfindexMin = 1 << 30
+	syntheticInterfaceIfindexMax = syntheticInterfaceIfindexMin + (1 << 20) - 1
+)
+
+func syntheticLogicalIfindex(name string, vlanID int, used map[int]struct{}) int {
+	if used == nil {
+		used = make(map[int]struct{})
+	}
+	const fnvOffset = uint32(2166136261)
+	const fnvPrime = uint32(16777619)
+	hash := fnvOffset
+	for _, b := range []byte(fmt.Sprintf("%s/%d", name, vlanID)) {
+		hash ^= uint32(b)
+		hash *= fnvPrime
+	}
+	span := syntheticInterfaceIfindexMax - syntheticInterfaceIfindexMin + 1
+	start := syntheticInterfaceIfindexMin + int(hash%uint32(span))
+	for offset := 0; offset < span; offset++ {
+		candidate := syntheticInterfaceIfindexMin + ((start - syntheticInterfaceIfindexMin + offset) % span)
+		if _, exists := used[candidate]; exists {
+			continue
+		}
+		used[candidate] = struct{}{}
+		return candidate
+	}
+	panic(fmt.Sprintf(
+		"userspace snapshot: exhausted synthetic ifindex range for %q (vlan=%d hash=%d tried=%d range=[%d,%d]); check for hash collisions or excessive logical-only VLAN units",
+		name,
+		vlanID,
+		hash,
+		span,
+		syntheticInterfaceIfindexMin,
+		syntheticInterfaceIfindexMax,
+	))
+}
+
+func shouldUseLogicalOnlyParentBoundRethVLAN(cfg *config.Config, ifName string, unit *config.InterfaceUnit, childIfindex int, parentIfindex int) bool {
+	if cfg == nil || unit == nil || childIfindex > 0 || parentIfindex <= 0 || unit.VlanID <= 0 {
+		return false
+	}
+	if !strings.HasPrefix(ifName, "reth") {
+		return false
+	}
+	return config.LinuxIfName(cfg.ResolveReth(ifName)) != config.LinuxIfName(ifName)
+}
+
+// UserspaceBoundLinuxInterfaces returns the deduplicated, sorted set of
+// Linux interface names that the userspace dataplane will bind AF_XDP
+// sockets to for the given compiled config. This is the authoritative
+// allowlist used by the D3 RSS indirection path (#797) so that we only
+// reshape RSS on interfaces we actually steer into AF_XDP workers —
+// siblings like a spare mlx5 PF or a management netdev must not be
+// touched.
+//
+// Scope mirrors buildUserspaceIngressIfindexes() and
+// userspaceSkipsIngressInterface(): include zoned non-tunnel interfaces
+// excluding fxp*, em*, fab*, lo0, mgmt/control zones, and RETH member
+// children; plus every fabric's parent member (fab0/fab1 themselves are
+// IPVLAN overlays and are excluded above, but their physical parent is
+// where AF_XDP binds). For zoned VLAN units whose parent is the physical
+// interface, we emit the parent Linux name — that is the netdev the
+// AF_XDP socket actually binds to.
+//
+// Returns nil on nil config. Never returns an error: this is a
+// best-effort derivation used to scope a best-effort optimization.
+func UserspaceBoundLinuxInterfaces(cfg *config.Config) []string {
+	if cfg == nil {
+		return nil
+	}
+	ucfg := deriveUserspaceConfig(cfg)
+	// Build a snapshot without depending on ifindex resolution — the
+	// allowlist is by Linux name (what `ethtool` consumes), so ifindex
+	// lookups are unnecessary here. We reuse the shared filter via the
+	// real builder to stay in lock-step with binding logic.
+	snap := buildSnapshot(cfg, ucfg, 0, 0)
+	if snap == nil {
+		return nil
+	}
+	seen := make(map[string]struct{})
+	out := make([]string, 0)
+	add := func(name string) {
+		if name == "" {
+			return
+		}
+		if _, ok := seen[name]; ok {
+			return
+		}
+		seen[name] = struct{}{}
+		out = append(out, name)
+	}
+	for _, iface := range snap.Interfaces {
+		if iface.Zone == "" || userspaceSkipsIngressInterface(iface) {
+			continue
+		}
+		// Prefer the parent Linux name when present (VLAN units bind on
+		// the parent physical netdev); otherwise the iface's own name.
+		if iface.ParentLinuxName != "" {
+			add(iface.ParentLinuxName)
+		} else {
+			add(iface.LinuxName)
+		}
+	}
+	for _, fab := range snap.Fabrics {
+		add(fab.ParentLinuxName)
+	}
+	sort.Strings(out)
+	return out
+}
+
+func buildInterfaceSnapshots(cfg *config.Config) []InterfaceSnapshot {
+	if cfg == nil || len(cfg.Interfaces.Interfaces) == 0 {
+		return nil
+	}
+	zoneByInterface := buildInterfaceZoneMap(cfg)
+	usedSyntheticIfindexes := make(map[int]struct{})
+	// Build RETH RG lookup: physical member → RETH's RedundancyGroup.
+	// Physical members have RedundantParent set but RedundancyGroup=0;
+	// the RG is on the RETH. Without this, flow cache HA checks on
+	// RETH member egress interfaces return owner_rg=0 and bypass
+	// HA active/inactive validation, causing stale forwarding after failover.
+	rethRG := make(map[string]int)
+	for _, ifc := range cfg.Interfaces.Interfaces {
+		if ifc != nil && ifc.RedundantParent != "" {
+			if reth := cfg.Interfaces.Interfaces[ifc.RedundantParent]; reth != nil && reth.RedundancyGroup > 0 {
+				rethRG[ifc.Name] = reth.RedundancyGroup
+			}
+		}
+	}
+	names := make([]string, 0, len(cfg.Interfaces.Interfaces))
+	for name := range cfg.Interfaces.Interfaces {
+		names = append(names, name)
+	}
+	sort.Strings(names)
+	out := make([]InterfaceSnapshot, 0, len(names))
+	for _, name := range names {
+		iface := cfg.Interfaces.Interfaces[name]
+		if iface == nil {
+			continue
+		}
+		linuxName := snapshotLinuxName(cfg, name, iface, nil)
+		ifindex, mtu, hardwareAddr, addresses := buildLinkSnapshot(linuxName)
+		// Use the interface's own RG, or inherit from RETH parent.
+		rg := iface.RedundancyGroup
+		if rg <= 0 {
+			rg = rethRG[name]
+		}
+		out = append(out, InterfaceSnapshot{
+			Name:            name,
+			Zone:            zoneByInterface[name],
+			LinuxName:       linuxName,
+			ParentLinuxName: "",
+			Ifindex:         ifindex,
+			ParentIfindex:   0,
+			RXQueues:        userspaceRXQueueCount(linuxName),
+			VLANID:          0,
+			LocalFabric:     iface.LocalFabricMember,
+			RedundancyGroup: rg,
+			UnitCount:       len(iface.Units),
+			Tunnel:          iface.Tunnel != nil,
+			MTU:             mtu,
+			HardwareAddr:    hardwareAddr,
+			Addresses:       addresses,
+		})
+		if len(iface.Units) == 0 {
+			continue
+		}
+		unitNums := make([]int, 0, len(iface.Units))
+		for unitNum := range iface.Units {
+			unitNums = append(unitNums, unitNum)
+		}
+		sort.Ints(unitNums)
+		for _, unitNum := range unitNums {
+			unit := iface.Units[unitNum]
+			if unit == nil {
+				continue
+			}
+			var cosUnit *config.CoSInterfaceUnit
+			if cfg.ClassOfService != nil {
+				if cosIface := cfg.ClassOfService.Interfaces[name]; cosIface != nil {
+					cosUnit = cosIface.Units[unitNum]
+				}
+			}
+			unitName := fmt.Sprintf("%s.%d", name, unitNum)
+			parentLinux := snapshotLinuxName(cfg, name, iface, nil)
+			parentIfindex, parentMTU, parentHardwareAddr, _ := buildLinkSnapshot(parentLinux)
+			parentRXQueues := userspaceRXQueueCount(parentLinux)
+			linuxUnit := snapshotLinuxName(cfg, name, iface, unit)
+			ifindex, mtu, hardwareAddr, addresses := buildLinkSnapshot(linuxUnit)
+			rxQueues := userspaceRXQueueCount(linuxUnit)
+			logicalOnly := false
+			// Bondless RETH VLAN units can transmit through the parent
+			// AF_XDP netdev without a Linux VLAN child. Keep a unique
+			// logical ifindex for Rust FIB/filter/CoS state, while the
+			// existing ParentIfindex path remains the socket bind target.
+			if shouldUseLogicalOnlyParentBoundRethVLAN(cfg, name, unit, ifindex, parentIfindex) {
+				ifindex = syntheticLogicalIfindex(unitName, unit.VlanID, usedSyntheticIfindexes)
+				logicalOnly = true
+				if mtu == 0 {
+					mtu = parentMTU
+				}
+				if hardwareAddr == "" {
+					hardwareAddr = parentHardwareAddr
+				}
+				if rxQueues == 0 {
+					rxQueues = parentRXQueues
+				}
+			}
+			addresses = mergeInterfaceAddressSnapshots(addresses, buildConfiguredAddressSnapshots(unit.Addresses))
+			out = append(out, InterfaceSnapshot{
+				Name:                      unitName,
+				Zone:                      zoneByInterface[unitName],
+				LinuxName:                 linuxUnit,
+				ParentLinuxName:           parentLinux,
+				Ifindex:                   ifindex,
+				ParentIfindex:             parentIfindex,
+				LogicalOnly:               logicalOnly,
+				RXQueues:                  rxQueues,
+				VLANID:                    unit.VlanID,
+				LocalFabric:               iface.LocalFabricMember,
+				RedundancyGroup:           rg, // inherit resolved RG (RETH parent or own)
+				UnitCount:                 0,
+				Tunnel:                    iface.Tunnel != nil || unit.Tunnel != nil,
+				MTU:                       mtu,
+				HardwareAddr:              hardwareAddr,
+				Addresses:                 addresses,
+				FilterInputV4:             unit.FilterInputV4,
+				FilterOutputV4:            unit.FilterOutputV4,
+				FilterInputV6:             unit.FilterInputV6,
+				FilterOutputV6:            unit.FilterOutputV6,
+				CoSShapingRateBytesPerSec: coSUnitShapingRate(cosUnit),
+				CoSBurstSize:              coSUnitBurstSize(cosUnit),
+				CoSSchedulerMap:           coSUnitSchedulerMap(cosUnit),
+				CoSDSCPClassifier:         coSUnitDSCPClassifier(cosUnit),
+				CoSIEEE8021Classifier:     coSUnitIEEE8021Classifier(cosUnit),
+				CoSDSCPRewriteRule:        coSUnitDSCPRewriteRule(cosUnit),
+			})
+		}
+	}
+	return out
+}
+
+func coSUnitShapingRate(unit *config.CoSInterfaceUnit) uint64 {
+	if unit == nil {
+		return 0
+	}
+	return unit.ShapingRateBytes
+}
+
+func coSUnitBurstSize(unit *config.CoSInterfaceUnit) uint64 {
+	if unit == nil {
+		return 0
+	}
+	return unit.BurstSizeBytes
+}
+
+func coSUnitSchedulerMap(unit *config.CoSInterfaceUnit) string {
+	if unit == nil {
+		return ""
+	}
+	return unit.SchedulerMap
+}
+
+func coSUnitDSCPClassifier(unit *config.CoSInterfaceUnit) string {
+	if unit == nil {
+		return ""
+	}
+	return unit.DSCPClassifier
+}
+
+func coSUnitIEEE8021Classifier(unit *config.CoSInterfaceUnit) string {
+	if unit == nil {
+		return ""
+	}
+	return unit.IEEE8021Classifier
+}
+
+func coSUnitDSCPRewriteRule(unit *config.CoSInterfaceUnit) string {
+	if unit == nil {
+		return ""
+	}
+	return unit.DSCPRewriteRule
+}
+
+func snapshotLinuxName(cfg *config.Config, ifName string, iface *config.InterfaceConfig, unit *config.InterfaceUnit) string {
+	if iface == nil {
+		return config.LinuxIfName(ifName)
+	}
+	if unit != nil {
+		if tunnelNames := cfg.TunnelNameMap(); len(tunnelNames) > 0 {
+			ref := fmt.Sprintf("%s.%d", ifName, unit.Number)
+			if linuxName, ok := tunnelNames[ref]; ok && linuxName != "" {
+				return linuxName
+			}
+		}
+		if unit.VlanID > 0 {
+			return fmt.Sprintf("%s.%d", config.LinuxIfName(cfg.ResolveReth(ifName)), unit.VlanID)
+		}
+		if strings.HasPrefix(ifName, "reth") {
+			if unit.Number == 0 {
+				return config.LinuxIfName(cfg.ResolveReth(ifName))
+			}
+			return config.LinuxIfName(cfg.ResolveReth(fmt.Sprintf("%s.%d", ifName, unit.Number)))
+		}
+		if unit.Number == 0 {
+			return config.LinuxIfName(ifName)
+		}
+		return config.LinuxIfName(fmt.Sprintf("%s.%d", ifName, unit.Number))
+	}
+	if strings.HasPrefix(ifName, "reth") {
+		return config.LinuxIfName(cfg.ResolveReth(ifName))
+	}
+	return config.LinuxIfName(ifName)
+}
+
+func buildLinkSnapshot(linuxName string) (ifindex int, mtu int, hardwareAddr string, addresses []InterfaceAddressSnapshot) {
+	if linuxName == "" {
+		return 0, 0, "", nil
+	}
+	if link, err := net.InterfaceByName(linuxName); err == nil {
+		ifindex = link.Index
+	}
+	if link, err := netlink.LinkByName(linuxName); err == nil && link != nil {
+		mtu = link.Attrs().MTU
+		if hw := link.Attrs().HardwareAddr; len(hw) > 0 {
+			hardwareAddr = hw.String()
+		}
+		addresses = buildInterfaceAddressSnapshots(link)
+	}
+	return ifindex, mtu, hardwareAddr, addresses
+}
+
+func buildConfiguredAddressSnapshots(addrs []string) []InterfaceAddressSnapshot {
+	if len(addrs) == 0 {
+		return nil
+	}
+	out := make([]InterfaceAddressSnapshot, 0, len(addrs))
+	for _, cidr := range addrs {
+		ip, netw, err := net.ParseCIDR(cidr)
+		if err != nil || netw == nil {
+			continue
+		}
+		netw.IP = ip
+		family := "inet"
+		if ip.To4() == nil {
+			family = "inet6"
+		}
+		out = append(out, InterfaceAddressSnapshot{
+			Family:  family,
+			Address: netw.String(),
+			Scope:   int(netlink.SCOPE_UNIVERSE),
+		})
+	}
+	sort.Slice(out, func(i, j int) bool {
+		if out[i].Family != out[j].Family {
+			return out[i].Family < out[j].Family
+		}
+		return out[i].Address < out[j].Address
+	})
+	return out
+}
+
+func mergeInterfaceAddressSnapshots(live []InterfaceAddressSnapshot, configured []InterfaceAddressSnapshot) []InterfaceAddressSnapshot {
+	if len(live) == 0 {
+		return configured
+	}
+	if len(configured) == 0 {
+		return live
+	}
+	seen := make(map[string]bool, len(live)+len(configured))
+	out := make([]InterfaceAddressSnapshot, 0, len(live)+len(configured))
+	for _, addr := range live {
+		key := addr.Family + "/" + addr.Address
+		if seen[key] {
+			continue
+		}
+		seen[key] = true
+		out = append(out, addr)
+	}
+	for _, addr := range configured {
+		key := addr.Family + "/" + addr.Address
+		if seen[key] {
+			continue
+		}
+		seen[key] = true
+		out = append(out, addr)
+	}
+	sort.Slice(out, func(i, j int) bool {
+		if out[i].Family != out[j].Family {
+			return out[i].Family < out[j].Family
+		}
+		return out[i].Address < out[j].Address
+	})
+	return out
+}
+
+func buildInterfaceAddressSnapshots(link netlink.Link) []InterfaceAddressSnapshot {
+	if link == nil {
+		return nil
+	}
+	addrs, err := netlink.AddrList(link, netlink.FAMILY_ALL)
+	if err != nil || len(addrs) == 0 {
+		return nil
+	}
+	out := make([]InterfaceAddressSnapshot, 0, len(addrs))
+	for _, addr := range addrs {
+		if addr.IPNet == nil {
+			continue
+		}
+		family := "inet"
+		if addr.IPNet.IP.To4() == nil {
+			family = "inet6"
+		}
+		out = append(out, InterfaceAddressSnapshot{
+			Family:  family,
+			Address: addr.IPNet.String(),
+			Scope:   addr.Scope,
+		})
+	}
+	sort.Slice(out, func(i, j int) bool {
+		if out[i].Family != out[j].Family {
+			return out[i].Family < out[j].Family
+		}
+		return out[i].Address < out[j].Address
+	})
+	return out
+}
+
+func userspaceRXQueueCount(linuxName string) int {
+	if linuxName == "" {
+		return 0
+	}
+	entries, err := os.ReadDir(filepath.Join("/sys/class/net", linuxName, "queues"))
+	if err != nil {
+		return 0
+	}
+	count := 0
+	for _, entry := range entries {
+		if !entry.IsDir() {
+			continue
+		}
+		if name := entry.Name(); len(name) > 3 && name[:3] == "rx-" {
+			count++
+		}
+	}
+	return count
+}
