+++ pkg/dataplane/userspace/routes.go
@@ -1,0 +1,221 @@
+package userspace
+
+import (
+	"net"
+	"slices"
+	"sort"
+	"strings"
+	"syscall"
+
+	"github.com/psaab/xpf/pkg/config"
+	"github.com/vishvananda/netlink"
+)
+
+func buildRouteSnapshots(cfg *config.Config, interfaces []InterfaceSnapshot) []RouteSnapshot {
+	if cfg == nil {
+		return nil
+	}
+	out := make([]RouteSnapshot, 0)
+	seen := make(map[string]struct{})
+	addSnapshot := func(snap RouteSnapshot) {
+		key := snap.Table + "|" + snap.Family + "|" + snap.Destination + "|" + strings.Join(snap.NextHops, ",") + "|" + snap.NextTable
+		if _, ok := seen[key]; ok {
+			return
+		}
+		seen[key] = struct{}{}
+		out = append(out, snap)
+	}
+	addRoutes := func(table, family string, routes []*config.StaticRoute) {
+		for _, route := range routes {
+			if route == nil {
+				continue
+			}
+			tableName, familyName := normalizeRouteSnapshotFamily(table, family, route.Destination)
+			snap := RouteSnapshot{
+				Table:       tableName,
+				Family:      familyName,
+				Destination: route.Destination,
+				Discard:     route.Discard,
+				NextTable:   route.NextTable,
+			}
+			for _, nh := range route.NextHops {
+				switch {
+				case nh.Address != "" && nh.Interface != "":
+					snap.NextHops = append(snap.NextHops, nh.Address+"@"+nh.Interface)
+				case nh.Address != "":
+					snap.NextHops = append(snap.NextHops, nh.Address)
+				case nh.Interface != "":
+					snap.NextHops = append(snap.NextHops, "@"+nh.Interface)
+				}
+			}
+			addSnapshot(snap)
+		}
+	}
+	interfaceTablesV4, interfaceTablesV6 := buildInterfaceRouteTables(cfg)
+	addConnectedRoutes := func(family, table string, prefixes []string) {
+		for _, prefix := range prefixes {
+			snap := RouteSnapshot{
+				Table:       table,
+				Family:      family,
+				Destination: prefix,
+			}
+			addSnapshot(snap)
+		}
+	}
+	addRoutes("inet.0", "inet", cfg.RoutingOptions.StaticRoutes)
+	addRoutes("inet6.0", "inet6", cfg.RoutingOptions.Inet6StaticRoutes)
+
+	if len(cfg.RoutingInstances) > 0 {
+		insts := make([]*config.RoutingInstanceConfig, 0, len(cfg.RoutingInstances))
+		for _, ri := range cfg.RoutingInstances {
+			if ri != nil {
+				insts = append(insts, ri)
+			}
+		}
+		sort.Slice(insts, func(i, j int) bool { return insts[i].Name < insts[j].Name })
+		for _, ri := range insts {
+			addRoutes(ri.Name+".inet.0", "inet", ri.StaticRoutes)
+			addRoutes(ri.Name+".inet6.0", "inet6", ri.Inet6StaticRoutes)
+		}
+	}
+	for _, iface := range interfaces {
+		if iface.Name == "" {
+			continue
+		}
+		v4Table := interfaceTablesV4[iface.Name]
+		if v4Table == "" {
+			v4Table = "inet.0"
+		}
+		v6Table := interfaceTablesV6[iface.Name]
+		if v6Table == "" {
+			v6Table = "inet6.0"
+		}
+		v4Prefixes, v6Prefixes := connectedPrefixesForInterface(iface)
+		addConnectedRoutes("inet", v4Table, v4Prefixes)
+		addConnectedRoutes("inet6", v6Table, v6Prefixes)
+	}
+
+	// Add synthetic routes for ip rule entries that implement inter-VRF
+	// route leaking (rib-groups, next-table). These rules send traffic
+	// matching a destination prefix to a different routing table.
+	// Without these, the userspace FIB can't cross-reference VRF tables.
+	tableIDToName := make(map[int]string)
+	for _, inst := range cfg.RoutingInstances {
+		if inst != nil && inst.TableID > 0 {
+			tableIDToName[inst.TableID] = inst.Name + ".inet.0"
+		}
+	}
+	for _, family := range []int{syscall.AF_INET, syscall.AF_INET6} {
+		rules, err := netlink.RuleList(family)
+		if err != nil {
+			continue
+		}
+		for _, rule := range rules {
+			if rule.Dst == nil || rule.Table <= 0 {
+				continue
+			}
+			tableName, ok := tableIDToName[rule.Table]
+			if !ok {
+				continue
+			}
+			familyStr := "inet"
+			mainTable := "inet.0"
+			if family == syscall.AF_INET6 {
+				familyStr = "inet6"
+				mainTable = "inet6.0"
+			}
+			addSnapshot(RouteSnapshot{
+				Table:       mainTable,
+				Family:      familyStr,
+				Destination: rule.Dst.String(),
+				NextTable:   tableName,
+			})
+		}
+	}
+
+	sort.Slice(out, func(i, j int) bool {
+		if out[i].Table != out[j].Table {
+			return out[i].Table < out[j].Table
+		}
+		if out[i].Family != out[j].Family {
+			return out[i].Family < out[j].Family
+		}
+		return out[i].Destination < out[j].Destination
+	})
+	return out
+}
+
+func buildInterfaceRouteTables(cfg *config.Config) (map[string]string, map[string]string) {
+	v4 := make(map[string]string)
+	v6 := make(map[string]string)
+	if cfg == nil {
+		return v4, v6
+	}
+	for _, ri := range cfg.RoutingInstances {
+		if ri == nil || ri.Name == "" {
+			continue
+		}
+		for _, ifname := range ri.Interfaces {
+			if ifname == "" {
+				continue
+			}
+			v4[ifname] = ri.Name + ".inet.0"
+			v6[ifname] = ri.Name + ".inet6.0"
+		}
+	}
+	return v4, v6
+}
+
+func connectedPrefixesForInterface(iface InterfaceSnapshot) ([]string, []string) {
+	var v4 []string
+	var v6 []string
+	for _, addr := range iface.Addresses {
+		if addr.Scope != 0 && addr.Scope != int(netlink.SCOPE_UNIVERSE) {
+			continue
+		}
+		ip, network, err := net.ParseCIDR(addr.Address)
+		if err != nil || network == nil {
+			continue
+		}
+		ones, bits := network.Mask.Size()
+		if ones <= 0 || ones == bits {
+			continue
+		}
+		network.IP = ip.Mask(network.Mask)
+		prefix := network.String()
+		switch addr.Family {
+		case "inet":
+			v4 = append(v4, prefix)
+		case "inet6":
+			if ip.IsLinkLocalUnicast() {
+				continue
+			}
+			v6 = append(v6, prefix)
+		}
+	}
+	slices.Sort(v4)
+	slices.Sort(v6)
+	return slices.Compact(v4), slices.Compact(v6)
+}
+
+func normalizeRouteSnapshotFamily(table, family, destination string) (string, string) {
+	isIPv6 := strings.Contains(destination, ":")
+	if isIPv6 {
+		family = "inet6"
+		switch {
+		case table == "inet.0":
+			table = "inet6.0"
+		case strings.HasSuffix(table, ".inet.0"):
+			table = strings.TrimSuffix(table, ".inet.0") + ".inet6.0"
+		}
+		return table, family
+	}
+	family = "inet"
+	switch {
+	case table == "inet6.0":
+		table = "inet.0"
+	case strings.HasSuffix(table, ".inet6.0"):
+		table = strings.TrimSuffix(table, ".inet6.0") + ".inet.0"
+	}
+	return table, family
+}
