+++ pkg/dataplane/userspace/zones.go
@@ -1,0 +1,71 @@
+package userspace
+
+import (
+	"fmt"
+	"sort"
+	"strings"
+
+	"github.com/psaab/xpf/pkg/config"
+)
+
+func buildZoneSnapshots(cfg *config.Config) []ZoneSnapshot {
+	if cfg == nil || len(cfg.Security.Zones) == 0 {
+		return nil
+	}
+	names := make([]string, 0, len(cfg.Security.Zones))
+	for name := range cfg.Security.Zones {
+		names = append(names, name)
+	}
+	sort.Strings(names)
+	out := make([]ZoneSnapshot, 0, len(names))
+	for i, name := range names {
+		out = append(out, ZoneSnapshot{
+			Name: name,
+			ID:   uint16(i + 1),
+		})
+	}
+	return out
+}
+
+func buildInterfaceZoneMap(cfg *config.Config) map[string]string {
+	if cfg == nil || len(cfg.Security.Zones) == 0 {
+		return nil
+	}
+	out := make(map[string]string, len(cfg.Security.Zones))
+	zoneNames := make([]string, 0, len(cfg.Security.Zones))
+	for name := range cfg.Security.Zones {
+		zoneNames = append(zoneNames, name)
+	}
+	sort.Strings(zoneNames)
+	for _, zoneName := range zoneNames {
+		zone := cfg.Security.Zones[zoneName]
+		if zone == nil {
+			continue
+		}
+		for _, iface := range zone.Interfaces {
+			if iface == "" {
+				continue
+			}
+			if _, exists := out[iface]; !exists {
+				out[iface] = zoneName
+			}
+			if base, unit, ok := strings.Cut(iface, "."); ok && base != "" {
+				if _, exists := out[base]; !exists {
+					out[base] = zoneName
+				}
+				if unit != "" {
+					continue
+				}
+			}
+			if ifCfg := cfg.Interfaces.Interfaces[iface]; ifCfg != nil {
+				for unitNum := range ifCfg.Units {
+					unitName := fmt.Sprintf("%s.%d", iface, unitNum)
+					if _, exists := out[unitName]; !exists {
+						out[unitName] = zoneName
+					}
+				}
+			}
+		}
+	}
+	return out
+}
